Exclude internal transactions that represent parent transaction’s value
This change causes the `transaction_hash_to_internal_transactions`
function to exclude an internal transaction from the results if it
matches the parent transaction’s value, to, and from address.

If there are multiple internal transactions that match on these
criteria, the one with the lowest id is excluded for consistent results.

Co-authored-by: pkananen <pkananen@gmail.com>