Store the Contract code in the Address
Includes a new column in Addresses table to store the
Contract bytecode. Having this code means that an Adress is a
contract.

We're gathering the attributes (`created_contract_code` and
`created_contract_address_hash`) that represent a Contract creation
in the Address fetching (during the indexing) so an Address can be
created in the database with them.