fetched_balance_block_number instead of fetched_balance_at
Fixes #243

Instead of getting the address balance at the latest block and recording
the timestamp of the insert, which is only loosely correlated with even
the latest block's timestamp on-chain, use the block number for the last
time an address was used in the indexed address foreign keys in the
chain:

* blocks
  * miner_hash
* internal_transactions
  * created_contract_address_hash
  * from_address_hash
  * to_address_hash
* logs
  * address_hash
* transactions
  * from_address_hash
  * to_address_hash

To gather `fetched_balance_block_number` on reboot,
`Explorer.Chain.stream_unfetched_addresses` queries across the above
columns for the `MAX(blocks.number)`, which is used in
`Explorer.Indexer.AddressBalanceFetcher.init/2`.

During indexing,
`Explorer.Indexer.AddressBalanceFetcher.async_fetch_balances/1` now
requires a list of `%{block_number: Block.block_number(), hash:
Hash.Truncated.t()}` instead of list of `Hash.Truncated.t()`.

Having a `block_number` available for all importable entities's
addresses required changing `Explorer.Indexer.AddressExtraction` to
extract the `block_number` too.

The use of `insert` in the factories was making tests that show the
count not make sense, so the factories were switched to `build` because
it means only the exact count is created when a default value is
overridden.  This does, unfortunately, mean that `params_for(factory)`
no longer generates foreign keys and they must be set explicitly OR the
association passed instead of the foreign key.

Allow nil extracted fetched_balance_block_number when pending transaction

*NOTE*: This is a database drop and reindex change.