Use Mox by default for local integration tests
How to run without Mox is in the root `README.md`.

Run all combinations of Variant and Transport on CircleCI to ensure that
Mox matches HTTP behavior, so Mox can be used for fast local testing:
* Geth, HTTP
* Geth, Mox
* Parity, HTTP
* Parity, Mox

Using Mox locally also lessens load on Sokol and the Infura API key the
tests use for Geth under the assumption that we do far more local
testing than CI testing.  May help with #340, but does not fix the root
cause, which is now known as `Sequence` not chunking ranges correctly on
retry.