Adds API docs page with 'try it' UI
Why:

* We'd like to provide an API documentation page with a 'Try it out' UI
for the API. We first considered using Swagger for this but when we
looked into it we realized it didn't fit our needs. It doesn't fit our
needs because Swagger is an API specification and the Swagger UI is
built for this specification. The Etherscan-compatible API we're
building doesn't meet the Swagger specification (name was changed to
OpenAPI on the latest version) so we can't easily use Swagger tools to
document our API.
* Issue link: https://github.com/poanetwork/poa-explorer/issues/454

This change addresses the need by:

* Adding API documentation page with 'Try it out' UI at `/api_docs`. The
API documentation data is defined within `ExplorerWeb.Etherscan` and
rendered by Phoenix templates, with partial templates for 'module
cards', 'action tiles', and 'model tables'. The `Try it out` UI is
powered by the `try_api.js` file, which shows/hides the 'Try it out' UI
as needed and provides support for testing the API from the docs page.