Realtime fetcher retries for unknown block number
Why:

* We want to add a retry mechanism, instead of waiting half a second for
every block the realtime fetcher attempts to fetch and index. We had
this half a second wait in place in an attempt to avoid "Unknown block
number" errors. The half a second wait was effective at reducing these
errors but it also meant we were awaiting unnecessarily for most
fetches. Retrying after half a second, when Parity says it
doesn't yet know of a given block number, makes more sense.
* Issue link: n/a

This change addresses the need by:

* Editing `Indexer.Block.Realtime.Fetcher.fetch_and_import_block/3` to
retry failed fetch/import attempts because of "Unknown block number"
errors. It waits half a second before retrying and retries a maximum of
three times.