feat: synchronously fetch coin balances when an address is viewed.
This is only done if the latest balance we have is as of a block
that is older than approximately one day. We could just fetch a
day old block and compare, but using the average_block_time lets
us avoid hitting the database again, and also accounts for somehow
having bad information in the blocks table, getting us the latest
balance no matter what.

If there is an unfetched coin balance, then we just fetch that
instead of doing the above. We also don't render the balance
if there is an unfetched row.