Force block refetch if transaction is recollated in a different block
Due to race conditions described in #1911 transactions from a consensus
block might get overwritten by the same transactions from a non-consensus
block.

To prevent this we force a block refetch (by marking it as non-consensus),
if a transaction belonging to it gets overwritten by the same transaction
from a different block.